requirements.txt
fastapi
uvicorn
hume
python-multipart
python-dotenv
httpx

